factor unsafe_task into detail/unsafe_task.hpp
