+++ include/tmc/detail/compat.hpp
@@ -1,3 +1,8 @@
+// Copyright (c) 2023-2025 Logan McDougall
+//
+// Distributed under the Boost Software License, Version 1.0. (See accompanying
+// file LICENSE or copy at http://www.boost.org/LICENSE_1_0.txt)
+
 #pragma once
 
 #include <atomic>
