+++ include/tmc/detail/concepts.hpp
@@ -4,6 +4,7 @@
 // file LICENSE or copy at http://www.boost.org/LICENSE_1_0.txt)
 
 #pragma once
+
 #include <optional>
 #include <type_traits>
